% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxima-package.R
\docType{package}
\name{maxima-package}
\alias{maxima}
\alias{maxima-package}
\title{maxima: Embed the 'Maxima' Computer Algebra System via 'ECL'}
\description{
Embeds the 'Maxima' computer algebra system directly into R by linking a vendored, compiled 'ECL' (Embeddable Common Lisp) runtime with a vendored, compiled 'Maxima' core, avoiding any subprocess or text-protocol overhead. Calls into 'Maxima' are ordinary in-process function calls. Provides a thin, generic mechanism for calling any 'Maxima' function from R and converting between R language objects and 'Maxima' expressions, without imposing a curated high-level API.
}
\author{
\strong{Maintainer}: Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}

Authors:
\itemize{
  \item Mitchell O'Hara-Wild \email{mail@mitchelloharawild.com}
}

Other contributors:
\itemize{
  \item Maxima developers (Maxima computer algebra system, vendored and compiled at install time; see inst/COPYRIGHTS) [contributor, copyright holder]
  \item ECL developers (ECL (Embeddable Common Lisp) runtime, vendored and compiled at install time; see inst/COPYRIGHTS) [contributor, copyright holder]
}

}
\keyword{internal}
